+++ Patches/ItemRubberDuckPatch.cs
@@ -1,0 +1,26 @@
+using HarmonyLib;
+
+namespace DisableItemsInShop;
+
+[HarmonyPatch(typeof(ItemRubberDuck))]
+static class ItemRubberDuckPatch
+{
+    private static bool RubberDuckDisabled = DisableItemsInShop.ShouldDisable() && DisableItemsInShop.BoundConfig.RubberDuck.Value;
+
+    [HarmonyPatch("Start")]
+    [HarmonyPostfix]
+    private static void Start_Postfix(ItemRubberDuck __instance)
+    {
+        if (!RubberDuckDisabled) return;
+        DisableItemsInShop.Logger.LogDebug($"Disabled item usage: {__instance.name}");
+    }
+
+    [HarmonyPatch("Quack")]
+    [HarmonyPrefix]
+    private static bool Quack_Prefix(ItemRubberDuck __instance)
+    {
+        if (!RubberDuckDisabled) return true;
+
+        return false;
+    }
+}