+++ Thunderstore/CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.2.0
+
+- Added support to disable melees & rubber ducks
+- Extended & optimized the config for more flexibility
+
+**Note: You may need to change the configuration again to your liking.**
+
 ## v1.1.1
 
 - Changed the config so you can select where you want to disable items
@@ -5,7 +12,7 @@
 ## v1.1.0
 
 - Added the possibility to also disable items in the lobby
-- optimized & cleaned up code
+- Optimized & cleaned up code
 
 ## v1.0.0